+++ demo/frontend/src/index.js
@@ -1,0 +1,3 @@
+import { initAll } from 'govuk-frontend';
+
+initAll();